all changes

+++ processes.php
@@ -7,21 +7,20 @@
  */
 include "connectmysql.php";
 
-if(isset($_GET['add']))
+if(isset($_GET['add']) && $_GET['add']==1)
 {
-    if(isset($_POST['addpname'])&&isset($_POST['addprice'])&&isset($_POST['addptype'])&&isset($_POST['addquantity'])&&isset($_POST['addpdesc'])&& is_numeric($_POST['addprice']))
+    if(isset($_POST['addpname'])&&isset($_POST['addprice'])&&isset($_POST['addptype'])&&isset($_POST['addquantity'])&&isset($_POST['addpdesc'])&& is_numeric($_POST['addprice']) && is_numeric($_POST['addquantity']))
     {
         $addpname = sanitize_input($_POST['addpname']);
         $addprice = sanitize_input($_POST['addprice']);
         $addptype = sanitize_input($_POST['addptype']);
         $addquantity = sanitize_input($_POST['addquantity']);
         $addpdesc = sanitize_input($_POST['addpdesc']);
-
         
         $sql = "INSERT INTO product_table (product_name, product_description, product_price, product_quantity, product_type)"
                 . " VALUES ('$addpname', '$addpdesc',' $addprice', '$addquantity', '$addptype')";
         $result = $conn->query($sql);
-        if($result == TRUE)
+        if($result != TRUE)
         {
             header("location:manageInventory");
         }
@@ -39,16 +38,25 @@ if(isset($_GET['add']))
     }
 }
 
-if(isset($_GET['remove']))
+$sql_order = 'SELECT * FROM p5_6.product_table';
+$result_order = $conn -> query($sql_order);
+
+if($result_order -> num_rows > 0)
 {
     $identifier = $_GET['remove'];
-    $sql = "DELETE FROM product_table WHERE product_name='$identifier'";
-    $result = $conn->query($sql);
-    header("location:manageInventory"); 
+    while($row_order = $result_order -> fetch_assoc())
+    {
+        if(isset($_GET['remove']) && !empty($_GET['remove']) && $row_order["product_name"] == $identifier)
+        {
+            $sql = "DELETE FROM product_table WHERE product_name='$identifier'";
+            $result = $conn->query($sql);
+            header("location:manageInventory"); 
+        }
+    }
 }
 
 
-if(isset($_GET['reset']))
+if($_GET['reset']==1)
 {
     session_start();
     unset($_SESSION["user"]);
@@ -64,10 +72,10 @@ if(isset($_POST['updateStatus']))
 {
     $elementId = sanitize_input($_POST['elementId']);
     $elementValue = sanitize_input($_POST['elementValue']);
-    $sql = "UPDATE p5_6.product_purchased SET delivery_status='$elementValue' WHERE product_id_purchased ='$elementId' ";
+    $sql = "UPDATE p5_6.product_purchased SET delivery_status='$elementValue' WHERE order_id ='$elementId' ";
     // Execute the query
     $result = $conn->query($sql);
-    header("location: aOrderTracking");
+    die;
 }
